+++ crates/egui/src/lib.rs
@@ -406,6 +406,7 @@
 #![allow(clippy::manual_range_contains)]
 
 mod animation_manager;
+mod atomics;
 pub mod cache;
 pub mod containers;
 mod context;
@@ -429,6 +430,7 @@ pub mod os;
 mod painter;
 mod pass_state;
 pub(crate) mod placer;
+mod plugin;
 pub mod response;
 mod sense;
 pub mod style;
@@ -442,7 +444,6 @@ mod widget_rect;
 pub mod widget_text;
 pub mod widgets;
 
-mod atomics;
 #[cfg(feature = "callstack")]
 #[cfg(debug_assertions)]
 mod callstack;
@@ -482,7 +483,7 @@ pub mod text {
 pub use self::{
     atomics::*,
     containers::{menu::MenuBar, *},
-    context::{Context, Plugin, RepaintCause, RequestRepaintInfo},
+    context::{Context, RepaintCause, RequestRepaintInfo},
     data::{
         Key, UserData,
         input::*,
@@ -501,6 +502,7 @@ pub use self::{
     load::SizeHint,
     memory::{Memory, Options, Theme, ThemePreference},
     painter::Painter,
+    plugin::Plugin,
     response::{InnerResponse, Response},
     sense::Sense,
     style::{FontSelection, Spacing, Style, TextStyle, Visuals},
